requirements.txt
pyspark==3.4.1
python-dotenv==1.0.0
boto3==1.26.0
pytest==7.4.0
